Add multithreading: separate simulation and ui threads (+)
- Move ui to mod ui and simulation to mod simulation
- Add new seeds
