+++ seeds.md
@@ -77,3 +77,5 @@
   - `breaming_uncivic_unfellow`
   - `valentin_fleurit`: my name
   - `unterse_trench_assures`
+  - `ethrog_tactus_stolons`
+  - `atavism_priests_physique`
